EnemyHitスクリプトのHPをintからfloatに変更した

+++ 死神お仕事！/Assets/Shishido/script/Enemy Hit.cs
@@ -7,7 +7,7 @@ public class EnemyHit : MonoBehaviour
 {
     Rigidbody2D rbody;              //Rigidbody2D型の変数
 
-    public int HP_E = 3;    //敵の体力
+    public float HP_E = 3;    //敵の体力
     private bool inDamage;  //ダメージ中のフラグ
 
     bool ZERO = false;//死亡確認
